[MOV] sale_project: split sale_order.py
